+++ packages/notifi-react/lib/components/TargetListItem.tsx
@@ -1,50 +1,30 @@
-import clsx from 'clsx';
 import React from 'react';
 
-import { Icon, IconType } from '../assets/Icons';
-import {
-  FormTarget,
-  Target,
-  TargetInfo,
-  ToggleTarget,
-  useNotifiTargetContext,
-  useNotifiTenantConfigContext,
-} from '../context';
-import { useComponentPosition } from '../hooks/useComponentPosition';
-import { useTargetWallet } from '../hooks/useTargetWallet';
-import {
-  getTargetValidateRegex,
-  hasValidTargetMoreThan,
-  isFormTarget,
-  isTargetCta,
-  isTargetVerified,
-  isToggleTarget,
-} from '../utils';
-import { PostCta, TargetCta, TargetCtaProps } from './TargetCta';
-import { TargetInputField } from './TargetInputField';
+import { IconType } from '../assets/Icons';
+import { FormTarget, TargetInfo, ToggleTarget } from '../context';
+import { isFormTarget, isToggleTarget } from '../utils';
+import { PostCta, TargetCtaProps } from './TargetCta';
+import { TargetListItemForm } from './TargetListItemForm';
+import { TargetListItemToggle } from './TargetListItemToggle';
 
-export type TargetListItemProps = {
+export type TargetListItemProps =
+  | TargetListItemToggleProps
+  | TargetListItemFromProps;
+export type TargetListItemToggleProps = TargetListItemPropsBase & {
+  target: ToggleTarget;
+};
+export type TargetListItemFromProps = TargetListItemPropsBase & {
+  target: FormTarget;
+};
+
+type TargetListItemPropsBase = {
   targetListRef: React.RefObject<HTMLDivElement>;
   postCta: PostCta;
   iconType: IconType;
   label: string;
   targetCtaType: TargetCtaProps['type'];
-  target: Target;
   targetInfo?: TargetInfo;
-  message?: {
-    beforeVerify?: string;
-    afterVerify?: string;
-    beforeVerifyTooltip?: string;
-    beforeVerifyTooltipEndingLink?: {
-      text: string;
-      url: string;
-    };
-    afterVerifyTooltip?: string;
-    afterVerifyTooltipEndingLink?: {
-      text: string;
-      url: string;
-    };
-  };
+  message?: TargetListItemMessage;
   parentComponent?: 'inbox' | 'ftu';
   classNames?: {
     targetListItem?: string;
@@ -60,444 +40,30 @@ export type TargetListItemProps = {
   };
 };
 
+export type TargetListItemMessage = {
+  beforeSignup?: string;
+  beforeSignupTooltip?: string;
+  beforeSignupTooltipEndingLink?: TooltipEndingLink;
+  beforeVerify?: string;
+  afterVerify?: string;
+  beforeVerifyTooltip?: string;
+  beforeVerifyTooltipEndingLink?: TooltipEndingLink;
+  afterVerifyTooltip?: string;
+  afterVerifyTooltipEndingLink?: TooltipEndingLink;
+};
+
+export type TooltipEndingLink = {
+  text: string;
+  url: string;
+};
+
 export const TargetListItem: React.FC<TargetListItemProps> = (props) => {
-  const tooltipRef = React.useRef<HTMLDivElement>(null);
-  const {
-    targetDocument: { targetData, targetInputs },
-    renewTargetGroup,
-    updateTargetInputs,
-  } = useNotifiTargetContext();
-  const { cardConfig } = useNotifiTenantConfigContext();
-  const {
-    signCoinbaseSignature,
-    isLoading: isLoadingWallet,
-    error: errorWallet,
-  } = useTargetWallet();
-
-  const isRemoveButtonAvailable = () => {
-    // TODO: Only for reference, remove before merge.
-    // const isRemoveButtonAvailable = (targetInfoPrompt: TargetInfoPrompt) => {
-    // if (cardConfig?.isContactInfoRequired) {
-    //   return (
-    //     getAvailableTargetInputCount(targetInputs) > 1 &&
-    //     // isTargetVerified(targetInfoPrompt) &&
-    //     props.parentComponent !== 'ftu'
-    //   );
-    // }
-    // return (
-    //   isTargetVerified(targetInfoPrompt) && props.parentComponent !== 'ftu'
-    // );
-    const isTargetRemovable = !!cardConfig?.isContactInfoRequired
-      ? hasValidTargetMoreThan(targetData, 1)
-      : true;
-
-    switch (props.target) {
-      case 'discord':
-        return (
-          !!props.targetInfo &&
-          props.targetInfo.infoPrompt.message !== 'Set up' &&
-          isTargetRemovable
-        );
-      case 'slack':
-        return (
-          !!props.targetInfo &&
-          props.targetInfo.infoPrompt.message !== 'Set Up' &&
-          isTargetRemovable
-        );
-      case 'wallet':
-        return (
-          !!props.targetInfo &&
-          props.targetInfo.infoPrompt.message !== 'Sign Wallet' &&
-          isTargetRemovable
-        );
-      default:
-        return !!props.targetInfo && isTargetRemovable;
-    }
-  };
-
-  const signupCtaProps: TargetCtaProps = React.useMemo(() => {
-    const defaultCtaProps: TargetCtaProps = {
-      type: 'button',
-      targetInfoPrompt: {
-        type: 'cta',
-        message: 'Signup',
-        onClick: async () => console.log('Default Signup placeHolder'),
-      },
-      postCta: props.postCta,
-      classNames: props.classNames?.TargetCta,
-    };
-
-    switch (props.target) {
-      case 'email':
-      case 'telegram':
-        return {
-          ...defaultCtaProps,
-          type: 'button',
-          targetInfoPrompt: {
-            type: 'cta',
-            message: 'Signup',
-            onClick: async () => {
-              const target = props.target as FormTarget;
-              renewTargetGroup({
-                target: target,
-                value: targetInputs[target].value,
-              });
-            },
-          },
-        };
-      case 'discord':
-        return {
-          ...defaultCtaProps,
-          isCtaDisabled: !targetData[props.target].isAvailable,
-          type: 'button',
-          targetInfoPrompt: {
-            type: 'cta',
-            message: 'Enable Bot',
-            onClick: async () => {
-              // TODO: Remove this after adding documentation: 1. single target subscription always sync with with targetData. 2. targetInput & multiple target subscription.
-              // await updateTargetInputs(props.target, true);
-              const targetGroup = await renewTargetGroup({
-                target: props.target as ToggleTarget,
-                value: true,
-              });
-
-              if (
-                targetGroup?.discordTargets?.[0]?.verificationLink &&
-                !targetGroup?.discordTargets?.[0]?.isConfirmed
-              ) {
-                window.open(
-                  targetGroup?.discordTargets?.[0]?.verificationLink,
-                  '_blank',
-                );
-              }
-            },
-          },
-        };
-      case 'wallet':
-        return {
-          ...defaultCtaProps,
-          isCtaDisabled: !targetData[props.target].isAvailable,
-          type: 'button',
-          targetInfoPrompt: {
-            type: 'cta',
-            message: 'Sign Wallet',
-            onClick: async () => {
-              // TODO: Remove this after adding documentation: 1. single target subscription always sync with with targetData. 2. targetInput & multiple target subscription.
-              // await updateTargetInputs(props.target, true);
-              const targetGroup = await renewTargetGroup({
-                target: props.target as ToggleTarget,
-                value: true,
-              });
-              // TODO: Handle error
-              const walletTargetId = targetGroup?.web3Targets?.[0]?.id;
-              const walletTargetSenderAddress =
-                targetGroup?.web3Targets?.[0]?.senderAddress;
-              if (
-                !targetGroup?.web3Targets?.[0]?.isConfirmed &&
-                walletTargetId &&
-                walletTargetSenderAddress
-              ) {
-                // TODO: Remove unused variable
-                const updatedWeb3Target = await signCoinbaseSignature(
-                  walletTargetId,
-                  walletTargetSenderAddress,
-                );
-              }
-            },
-          },
-        };
-      case 'slack':
-        return {
-          ...defaultCtaProps,
-          isCtaDisabled: !targetData[props.target].isAvailable,
-          type: 'button',
-          targetInfoPrompt: {
-            type: 'cta',
-            message: 'Signup',
-            onClick: async () => {
-              await updateTargetInputs(props.target, true);
-              const targetGroup = await renewTargetGroup({
-                target: props.target as ToggleTarget,
-                value: true,
-              });
-              const verificationLink =
-                targetGroup?.slackChannelTargets?.[0]?.verificationLink;
-              if (!verificationLink) return;
-              window.open(verificationLink, '_blank');
-            },
-          },
-        };
-      default:
-        return defaultCtaProps;
-    }
-  }, [
-    props.target,
-    targetInputs /* renewTargetGroup, updateTargetInputs, signCoinbaseSignature */, //TODO: econsider the dependency
-  ]);
-
-  const { componentPosition: tooltipIconPosition } = useComponentPosition(
-    tooltipRef,
-    props.parentComponent === 'inbox'
-      ? 'notifi-inbox-config-target-list-main'
-      : 'notifi-ftu-target-list-main',
-  );
-
   if (isFormTarget(props.target))
-    return (
-      <div
-        className={clsx(
-          'notifi-target-list-item',
-          props.classNames?.targetListItem,
-          // NOTE: only used when we want to adopt different style for verified items
-          isTargetVerified(props.targetInfo?.infoPrompt) &&
-            props.classNames?.targetListVerifiedItem,
-        )}
-      >
-        <div
-          className={clsx(
-            'notifi-target-list-item-target',
-            props.classNames?.targetListItemTarget,
-          )}
-        >
-          <Icon
-            type={props.iconType}
-            className={clsx('notifi-target-list-icon', props.classNames?.icon)}
-          />
-          <div
-            className={clsx(
-              'notifi-target-list-item-target-id',
-              props.classNames?.targetId,
-            )}
-          >
-            {/** TODO: Move to use memo once the target display id > 1 format */}
-            {targetData[props.target]}
-          </div>
-          {/* TODO */}
-          {!props.targetInfo ? (
-            <>
-              <label>{props.label}</label>{' '}
-              <TargetInputField
-                targetType={props.target}
-                validateRegex={getTargetValidateRegex(props.target)}
-              />
-            </>
-          ) : null}
-        </div>
-        {/* TODO: impl after verify message for form targets */}
-        {props.message?.beforeVerify &&
-        isTargetCta(props.targetInfo?.infoPrompt) ? (
-          <div
-            className={clsx(
-              'notifi-target-list-target-verify-message',
-              props.classNames?.verifyMessage,
-            )}
-          >
-            {props.message.beforeVerify}
-          </div>
-        ) : null}
-
-        {props.targetInfo ? (
-          <TargetCta
-            type={props.targetCtaType}
-            targetInfoPrompt={props.targetInfo.infoPrompt}
-            classNames={props.classNames?.TargetCta}
-            postCta={props.postCta}
-          />
-        ) : (
-          <>
-            {!targetInputs[props.target].error &&
-            targetInputs[props.target].value ? (
-              <TargetCta {...signupCtaProps} />
-            ) : null}
-          </>
-        )}
-        {isRemoveButtonAvailable() ? (
-          <TargetListItemAction
-            action={async () => {
-              const target = props.target as FormTarget;
-              updateTargetInputs(target, { value: '' });
-              renewTargetGroup({
-                target: target,
-                value: '',
-              });
-            }}
-            classNames={{ removeCta: props.classNames?.removeCta }}
-          />
-        ) : null}
-      </div>
-    );
+    return <TargetListItemForm {...(props as TargetListItemFromProps)} />;
 
   if (isToggleTarget(props.target)) {
-    const toggleTargetData = targetData[props.target];
-    return (
-      <div
-        className={clsx(
-          'notifi-target-list-item',
-          props.classNames?.targetListItem,
-          // NOTE: only used when we want to adopt different style for verified items
-          isTargetVerified(props.targetInfo?.infoPrompt) &&
-            props.classNames?.targetListVerifiedItem,
-        )}
-      >
-        <div
-          className={clsx(
-            'notifi-target-list-item-target',
-            props.classNames?.targetListItemTarget,
-          )}
-        >
-          <Icon
-            type={props.iconType}
-            className={clsx('notifi-target-list-icon', props.classNames?.icon)}
-          />
-          <label>{props.label}</label>
-        </div>
-        {!isTargetVerified(props.targetInfo?.infoPrompt) ||
-        props.parentComponent === 'ftu' ||
-        props.target === 'wallet' ? null : (
-          <div
-            className={clsx(
-              'notifi-target-list-item-target-id',
-              props.classNames?.targetId,
-            )}
-          >
-            {/** TODO: Move to use memo once the target display id > 1 format */}
-            {/** Display Discord username */}
-            {toggleTargetData?.data &&
-              'username' in toggleTargetData.data &&
-              `@${toggleTargetData.data.username}`}
-          </div>
-        )}
-        {props.message?.beforeVerify &&
-        (isTargetCta(props.targetInfo?.infoPrompt) || !props.targetInfo) ? (
-          <div
-            className={clsx(
-              'notifi-target-list-target-verify-message',
-              props.classNames?.verifyMessage,
-            )}
-          >
-            {props.message.beforeVerify}
-            <div className={'notifi-target-list-item-tooltip'} ref={tooltipRef}>
-              <Icon
-                className={clsx(
-                  'notifi-target-list-item-tooltip-icon',
-                  props.classNames?.tooltipIcon,
-                )}
-                type="info"
-              />
-              <div
-                className={clsx(
-                  'notifi-target-list-item-tooltip-content',
-                  props.classNames?.tooltipContent,
-                  props.parentComponent === 'inbox' ? 'inbox' : '',
-                  tooltipIconPosition,
-                )}
-              >
-                {props.message.beforeVerifyTooltip}{' '}
-                {props.message.beforeVerifyTooltipEndingLink ? (
-                  <a
-                    href={props.message.beforeVerifyTooltipEndingLink.url}
-                    target="_blank"
-                    rel="noopener noreferrer"
-                  >
-                    {props.message.beforeVerifyTooltipEndingLink.text}
-                  </a>
-                ) : null}
-              </div>
-            </div>
-          </div>
-        ) : null}
-        {props.message?.afterVerify &&
-        isTargetVerified(props.targetInfo?.infoPrompt) ? (
-          <div
-            className={clsx(
-              'notifi-target-list-target-confirmed-message',
-              props.classNames?.verifyMessage,
-              props.parentComponent === 'inbox' ? 'inbox' : '',
-            )}
-          >
-            {props.message.afterVerify}
-            <div className={'notifi-target-list-item-tooltip'} ref={tooltipRef}>
-              <Icon
-                className={clsx(
-                  'notifi-target-list-item-tooltip-icon',
-                  props.classNames?.tooltipIcon,
-                )}
-                type="info"
-              />
-              <div
-                className={clsx(
-                  'notifi-target-list-item-tooltip-content',
-                  props.classNames?.tooltipContent,
-                  props.parentComponent === 'inbox' ? 'inbox' : '',
-                  tooltipIconPosition,
-                )}
-              >
-                {props.message.afterVerifyTooltip}{' '}
-                {props.message.afterVerifyTooltipEndingLink ? (
-                  <a
-                    href={props.message.afterVerifyTooltipEndingLink.url}
-                    target="_blank"
-                    rel="noopener noreferrer"
-                  >
-                    {props.message.afterVerifyTooltipEndingLink.text}
-                  </a>
-                ) : null}
-              </div>
-            </div>
-          </div>
-        ) : null}
-        {props.targetInfo ? (
-          <TargetCta
-            type={props.targetCtaType}
-            targetInfoPrompt={props.targetInfo.infoPrompt}
-            classNames={props.classNames?.TargetCta}
-            postCta={props.postCta}
-            isCtaDisabled={!targetData[props.target].isAvailable}
-          />
-        ) : (
-          <TargetCta {...signupCtaProps} />
-        )}
-
-        {isRemoveButtonAvailable() ? (
-          <TargetListItemAction
-            action={async () => {
-              // TODO: Remove this after adding documentation: 1. single target subscription always sync with with targetData. 2. targetInput & multiple target subscription.
-              // updateTargetInputs(props.target, false); // TODO: remove
-              renewTargetGroup({
-                target: props.target as ToggleTarget,
-                value: false,
-              });
-            }}
-            classNames={{ removeCta: props.classNames?.removeCta }}
-          />
-        ) : null}
-      </div>
-    );
+    return <TargetListItemToggle {...(props as TargetListItemToggleProps)} />;
   }
 
   return null;
-};
-
-type TargetListItemActionProps = {
-  action: () => Promise<void>;
-  classNames?: {
-    removeCta?: string;
-  };
-  actionText?: string;
-};
-
-export const TargetListItemAction: React.FC<TargetListItemActionProps> = (
-  props,
-) => {
-  return (
-    <div
-      className={clsx(
-        'notifi-target-list-item-remove',
-        props.classNames?.removeCta,
-      )}
-      onClick={props.action}
-    >
-      {props.actionText ?? 'Remove'}
-    </div>
-  );
 };
