+++ packages/notifi-react/lib/components/TargetListItemAction.tsx
@@ -1,0 +1,26 @@
+import clsx from 'clsx';
+import React from 'react';
+
+type TargetListItemActionProps = {
+  action: () => Promise<void>;
+  classNames?: {
+    removeCta?: string;
+  };
+  actionText?: string;
+};
+
+export const TargetListItemAction: React.FC<TargetListItemActionProps> = (
+  props,
+) => {
+  return (
+    <div
+      className={clsx(
+        'notifi-target-list-item-remove',
+        props.classNames?.removeCta,
+      )}
+      onClick={props.action}
+    >
+      {props.actionText ?? 'Remove'}
+    </div>
+  );
+};
